requirements.txt
asgiref==3.8.1
Django==5.1
django-environ==0.11.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-tenants==3.7.0